feat: #13305, data type is too small, can be overflow, thus some variables are declared as long long

+++ cote13305.cpp
@@ -3,25 +3,25 @@
 
 class solver_13305{
     public:
-    solver_13305(std::vector<long long int> distance, std::vector<long long int> cost){
+    solver_13305(std::vector<int> distance, std::vector<int> cost){
         distances = distance;
         costs = cost;
         N = costs.size();
     }
 
-    int solver(){
-        long long int total_cost = 0;
-        long long int now_cost = costs[0];
-        long long int next_cost;
+    long long solver(){
+        long long total_cost = 0;
+        int now_cost = costs[0];
+        int next_cost;
         for(int i =1 ; i < N ; ++i){
             next_cost = costs[i];
 
             if (now_cost > next_cost){
-                total_cost += distances[i-1] * now_cost;
+                total_cost += (long long)distances[i-1] * now_cost;
                 now_cost = costs[i];
             }
             else{
-                total_cost += distances[i-1] * now_cost;
+                total_cost += (long long)distances[i-1] * now_cost;
             }
         }
         
@@ -29,7 +29,7 @@ class solver_13305{
     }
 
     private:
-    std::vector<long long int> distances, costs;
+    std::vector<int> distances, costs;
     int N;
 
 };
@@ -39,26 +39,27 @@ int main(){
     int N; // 도시의 개수
     std::cin >> N;
 
-    long long int distance;
-    std::vector<long long int> distances;
+    int distance;
+    std::vector<int> distances;
     for(int i = 0 ; i < (N-1) ; ++i){
         std::cin >> distance;
         distances.push_back(distance);
     }
 
-    long long int cost;
-    std::vector<long long int> costs;
+    int cost;
+    std::vector<int> costs;
     for(int i =0 ; i < N ; ++i){
         std::cin >> cost;
         costs.push_back(cost);
     }
 
     solver_13305 solver(distances, costs);
-    long long int answer = solver.solver();
+    long long answer = solver.solver();
 
     std::cout << answer << std::endl;
     return 0;
 }
+
 
 /*
 일단은, 지금 도시의 가격보다 다음 도시의 가겨이 싸면, 딱 이동할 거리만큼만 주유를 하고,
